Changed PyCrypto to PyCryptodome
This commit removes the usage of PyCrypto and add support for
PyCryptodome. This is necessary as PyCrypto is not maintained anymore
and seems to have serious issues. PyCryptodome is an active fork of
PyCrypto

This is referenced in dlitz/pycrypto#173 and dlitz/pycrypto#176.

This is an answer to the suggestion from raymontag/keepass#72